requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
pypdf
python-docx
streamlit
python-dotenv
openai